[Keycloak] Allow NodePort with LoadBalancer
Similarly to Ingress-Nginx
(https://github.com/kubernetes/ingress-nginx/blob/master/charts/ingress-nginx/templates/controller-service.yaml#L40-L48)
allow to set NodePort when Service of type LoadBalancer is used.
The use case for this is to allow using specific node port when Network
Load Balancer is used in AWS setup and a specific port needs to be used
for a case where security groups need to be managed outside of
Kubernets.